Creating new initialization entrypoints
