+++ fabric/forgero-fabric-core/src/main/java/com/sigmundgranaas/forgero/fabric/ForgeroInitializer.java
@@ -1,98 +1,46 @@
 package com.sigmundgranaas.forgero.fabric;
 
-import static com.sigmundgranaas.forgero.core.identifier.Common.ELEMENT_SEPARATOR;
-import static com.sigmundgranaas.forgero.minecraft.common.block.assemblystation.AssemblyStationBlock.*;
-import static com.sigmundgranaas.forgero.minecraft.common.block.assemblystation.AssemblyStationScreenHandler.ASSEMBLY_STATION_SCREEN_HANDLER;
-import static com.sigmundgranaas.forgero.minecraft.common.entity.Entities.SOUL_ENTITY;
-
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.List;
 import java.util.Set;
-import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
-import com.google.gson.Gson;
-import com.google.gson.stream.JsonReader;
-import com.sigmundgranaas.forgero.core.Forgero;
 import com.sigmundgranaas.forgero.core.ForgeroStateRegistry;
-import com.sigmundgranaas.forgero.core.condition.Conditions;
-import com.sigmundgranaas.forgero.core.condition.LootCondition;
 import com.sigmundgranaas.forgero.core.configuration.ForgeroConfigurationLoader;
-import com.sigmundgranaas.forgero.core.property.AttributeType;
-import com.sigmundgranaas.forgero.core.property.active.ActivePropertyRegistry;
-import com.sigmundgranaas.forgero.core.property.active.VeinBreaking;
-import com.sigmundgranaas.forgero.core.registry.SoulLevelPropertyRegistry;
 import com.sigmundgranaas.forgero.core.resource.PipelineBuilder;
-import com.sigmundgranaas.forgero.core.resource.data.v2.data.ConditionData;
-import com.sigmundgranaas.forgero.core.resource.data.v2.data.LootEntryData;
-import com.sigmundgranaas.forgero.core.resource.data.v2.data.SoulLevelPropertyData;
-import com.sigmundgranaas.forgero.core.soul.SoulLevelPropertyDataProcessor;
-import com.sigmundgranaas.forgero.core.state.State;
-import com.sigmundgranaas.forgero.core.type.Type;
-import com.sigmundgranaas.forgero.fabric.command.CommandRegistry;
-import com.sigmundgranaas.forgero.fabric.item.StateToItemConverter;
-import com.sigmundgranaas.forgero.fabric.loot.TreasureInjector;
-import com.sigmundgranaas.forgero.fabric.registry.DefaultLevelProperties;
-import com.sigmundgranaas.forgero.fabric.registry.RecipeRegistry;
+import com.sigmundgranaas.forgero.fabric.api.entrypoint.ForgeroInitializedEntryPoint;
+import com.sigmundgranaas.forgero.fabric.api.entrypoint.ForgeroPreInitializationEntryPoint;
 import com.sigmundgranaas.forgero.fabric.registry.RegistryHandler;
-import com.sigmundgranaas.forgero.fabric.resources.ARRPGenerator;
 import com.sigmundgranaas.forgero.fabric.resources.FabricPackFinder;
-import com.sigmundgranaas.forgero.fabric.resources.dynamic.MaterialPartTagGenerator;
-import com.sigmundgranaas.forgero.fabric.resources.dynamic.PartToSchematicGenerator;
-import com.sigmundgranaas.forgero.fabric.resources.dynamic.PartTypeTagGenerator;
-import com.sigmundgranaas.forgero.fabric.resources.dynamic.RepairKitResourceGenerator;
-import com.sigmundgranaas.forgero.fabric.resources.dynamic.SchematicPartTagGenerator;
-import com.sigmundgranaas.forgero.minecraft.common.entity.Entities;
-import com.sigmundgranaas.forgero.minecraft.common.entity.SoulEntity;
-import com.sigmundgranaas.forgero.minecraft.common.item.Attributes;
-import com.sigmundgranaas.forgero.minecraft.common.item.DynamicItems;
-import com.sigmundgranaas.forgero.minecraft.common.loot.SingleLootEntry;
-import com.sigmundgranaas.forgero.minecraft.common.loot.function.LootFunctions;
-import com.sigmundgranaas.forgero.minecraft.common.property.handler.PatternBreaking;
-import com.sigmundgranaas.forgero.minecraft.common.property.handler.TaggedPatternBreaking;
-import com.sigmundgranaas.forgero.minecraft.common.resources.DisassemblyRecipeLoader;
 import com.sigmundgranaas.forgero.minecraft.common.service.StateService;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
-import net.minecraft.resource.Resource;
-import net.minecraft.resource.ResourceManager;
-import net.minecraft.resource.ResourceType;
-import net.minecraft.util.Identifier;
-import net.minecraft.util.InvalidIdentifierException;
-import net.minecraft.util.registry.Registry;
 
 import net.fabricmc.api.ModInitializer;
-import net.fabricmc.fabric.api.object.builder.v1.entity.FabricDefaultAttributeRegistry;
-import net.fabricmc.fabric.api.resource.ResourceManagerHelper;
-import net.fabricmc.fabric.api.resource.ResourceReloadListenerKeys;
-import net.fabricmc.fabric.api.resource.SimpleSynchronousResourceReloadListener;
 import net.fabricmc.loader.api.FabricLoader;
 import net.fabricmc.loader.api.ModContainer;
+import net.fabricmc.loader.api.entrypoint.EntrypointContainer;
 import net.fabricmc.loader.api.metadata.ModMetadata;
 
 
 public class ForgeroInitializer implements ModInitializer {
-
-
 	public static final String MOD_NAMESPACE = "forgero";
-	public static final Logger LOGGER = LogManager.getLogger(ForgeroInitializer.MOD_NAMESPACE);
+	private static final List<ForgeroInitializedEntryPoint> INITIALIZED_ENTRY_POINTS =
+			FabricLoader.getInstance().getEntrypointContainers("forgeroInitialized", ForgeroInitializedEntryPoint.class)
+					.stream()
+					.map(EntrypointContainer::getEntrypoint)
+					.toList();
 
+	private static final List<ForgeroPreInitializationEntryPoint> PRE_INITIALIZED_ENTRY_POINTS =
+			FabricLoader.getInstance().getEntrypointContainers("forgeroPreInitialization", ForgeroPreInitializationEntryPoint.class)
+					.stream()
+					.map(EntrypointContainer::getEntrypoint)
+					.toList();
 
 	@Override
 	public void onInitialize() {
-		ActivePropertyRegistry.register(new ActivePropertyRegistry.PropertyEntry(PatternBreaking.predicate, PatternBreaking.factory));
-		ActivePropertyRegistry.register(new ActivePropertyRegistry.PropertyEntry(TaggedPatternBreaking.predicate, TaggedPatternBreaking.factory));
-		ActivePropertyRegistry.register(new ActivePropertyRegistry.PropertyEntry(VeinBreaking.predicate, VeinBreaking.factory));
+		handlePreInitializedEntryPoints();
 
 		Set<String> availableDependencies = FabricLoader.getInstance().getAllMods().stream().map(ModContainer::getMetadata).map(ModMetadata::getId).collect(Collectors.toSet());
 
 		var configuration = ForgeroConfigurationLoader.load();
-		soulLevelPropertyReloader();
-		Entities.register();
-		FabricDefaultAttributeRegistry.register(SOUL_ENTITY, SoulEntity.createSoulEntities());
 
 		PipelineBuilder
 				.builder()
@@ -108,218 +56,15 @@ public class ForgeroInitializer implements ModInitializer {
 				.build()
 				.execute();
 
-		var handler = RegistryHandler.getHandler();
-		handler.acceptJob(this::registerBlocks);
-		handler.acceptJob(this::registerStates);
-		handler.acceptJob(DynamicItems::registerDynamicItems);
-		handler.acceptJob(this::registerTreasure);
-		handler.acceptJob(this::registerCommands);
-		handler.acceptJob(this::registerLevelPropertiesDefaults);
-		handler.acceptJob(LootFunctions::register);
-		handler.acceptJob(Attributes::register);
-		handler.acceptJob(this::disassemblyReloader);
-
-		handler.runJobs();
-		dataReloader();
-		lootConditionReloader();
-		lootInjectionReloader();
-
-		handler.acceptSyncedJob(this::registerAARPRecipes);
-		handler.acceptSyncedJob((service) -> registerRecipes());
-		handler.initialize();
+		StateService service = RegistryHandler.getHandler().initialize();
+		handleInitializedEntryPoints(service);
 	}
 
-	private void registerLevelPropertiesDefaults() {
-		DefaultLevelProperties.defaults().forEach(SoulLevelPropertyRegistry::register);
+	private void handleInitializedEntryPoints(StateService service) {
+		INITIALIZED_ENTRY_POINTS.forEach(entryPoint -> entryPoint.onInitialized(service));
 	}
 
-	private void registerBlocks() {
-		Registry.register(Registry.BLOCK, ASSEMBLY_STATION, ASSEMBLY_STATION_BLOCK);
-		Registry.register(Registry.ITEM, ASSEMBLY_STATION, ASSEMBLY_STATION_ITEM);
-		Registry.register(Registry.SCREEN_HANDLER, ASSEMBLY_STATION, ASSEMBLY_STATION_SCREEN_HANDLER);
-	}
-
-	private void registerAARPRecipes(StateService service) {
-		ARRPGenerator.register(new RepairKitResourceGenerator(ForgeroConfigurationLoader.configuration));
-		ARRPGenerator.register(PartToSchematicGenerator::new);
-		ARRPGenerator.register(MaterialPartTagGenerator::new);
-		ARRPGenerator.register(SchematicPartTagGenerator::new);
-		ARRPGenerator.register(PartTypeTagGenerator::new);
-		ARRPGenerator.generate(service);
-	}
-
-	private void dataReloader() {
-		ResourceManagerHelper.get(ResourceType.SERVER_DATA).registerReloadListener(new SimpleSynchronousResourceReloadListener() {
-			@Override
-			public void reload(ResourceManager manager) {
-				var config = ForgeroConfigurationLoader.load();
-				Set<String> availableDependencies = FabricLoader.getInstance().getAllMods().stream().map(ModContainer::getMetadata).map(ModMetadata::getId).collect(Collectors.toSet());
-				PipelineBuilder.builder()
-						.register(() -> config)
-						.register(FabricPackFinder.supplier())
-						.state(ForgeroStateRegistry.stateListener())
-						.register(availableDependencies)
-						.silent()
-						.build()
-						.execute();
-			}
-
-			@Override
-			public Identifier getFabricId() {
-				return new Identifier(ForgeroInitializer.MOD_NAMESPACE, "data");
-			}
-		});
-	}
-
-
-	private void disassemblyReloader() {
-		ResourceManagerHelper.get(ResourceType.SERVER_DATA).registerReloadListener(new SimpleSynchronousResourceReloadListener() {
-			@Override
-			public void reload(ResourceManager manager) {
-				DisassemblyRecipeLoader.reload(manager);
-			}
-
-			@Override
-			public Identifier getFabricId() {
-				return new Identifier(Forgero.NAMESPACE, "disassembly");
-			}
-		});
-	}
-
-	private void lootConditionReloader() {
-		ResourceManagerHelper.get(ResourceType.SERVER_DATA).registerReloadListener(new SimpleSynchronousResourceReloadListener() {
-			@Override
-			public void reload(ResourceManager manager) {
-				Conditions.INSTANCE.refresh();
-				Gson gson = new Gson();
-				for (Resource res : manager.findResources("conditions", path -> path.getPath().endsWith(".json")).values()) {
-					try (InputStream stream = res.getInputStream()) {
-						ConditionData data = gson.fromJson(new JsonReader(new InputStreamReader(stream)), ConditionData.class);
-						LootCondition.of(data).ifPresent(Conditions.INSTANCE::register);
-					} catch (Exception e) {
-						Forgero.LOGGER.error(e);
-					}
-				}
-			}
-
-			@Override
-			public Identifier getFabricId() {
-				return new Identifier(Forgero.NAMESPACE, "loot_condition");
-			}
-		});
-	}
-
-	private void lootInjectionReloader() {
-		ResourceManagerHelper.get(ResourceType.SERVER_DATA).registerReloadListener(new SimpleSynchronousResourceReloadListener() {
-			@Override
-			public void reload(ResourceManager manager) {
-				Gson gson = new Gson();
-				TreasureInjector injector = TreasureInjector.getInstance();
-				for (Resource res : manager.findResources("forgero_loot", path -> path.getPath().endsWith(".json")).values()) {
-					try (InputStream stream = res.getInputStream()) {
-						LootEntryData data = gson.fromJson(new JsonReader(new InputStreamReader(stream)), LootEntryData.class);
-						injector.registerEntry(data.id(), SingleLootEntry.of(data));
-					} catch (Exception e) {
-						Forgero.LOGGER.error(e);
-					}
-				}
-				TreasureInjector.getInstance().registerDynamicLoot();
-			}
-
-			@Override
-			public Identifier getFabricId() {
-				return ResourceReloadListenerKeys.LOOT_TABLES;
-			}
-		});
-	}
-
-	private void soulLevelPropertyReloader() {
-		ResourceManagerHelper.get(ResourceType.SERVER_DATA).registerReloadListener(new SimpleSynchronousResourceReloadListener() {
-			@Override
-			public void reload(ResourceManager manager) {
-				SoulLevelPropertyRegistry.refresh();
-				Gson gson = new Gson();
-				for (Resource res : manager.findResources("leveled_soul_properties", path -> path.getPath().endsWith(".json")).values()) {
-					try (InputStream stream = res.getInputStream()) {
-						SoulLevelPropertyData data = gson.fromJson(new JsonReader(new InputStreamReader(stream)), SoulLevelPropertyData.class);
-						SoulLevelPropertyRegistry.register(data.getId(), new SoulLevelPropertyDataProcessor(data));
-					} catch (Exception e) {
-						Forgero.LOGGER.error(e);
-					}
-				}
-			}
-
-			@Override
-			public Identifier getFabricId() {
-				return new Identifier(Forgero.NAMESPACE, "soul_level_property");
-			}
-		});
-	}
-
-	private void registerTreasure() {
-		TreasureInjector.getInstance().registerLoot();
-	}
-
-	private void registerCommands() {
-		new CommandRegistry().registerCommand();
-	}
-
-	private void registerStates() {
-		var sortingMap = new HashMap<String, Integer>();
-
-		ForgeroStateRegistry.STATES.all().stream().map(Supplier::get)
-				.filter(state -> !state.test(Type.WEAPON) && !state.test(Type.TOOL)).forEach(state -> sortingMap.compute(materialName(state), (key, value) -> value == null || rarity(state) > value ? rarity(state) : value));
-
-		ForgeroStateRegistry.CREATE_STATES.stream()
-				.filter(state -> !Registry.ITEM.containsId(new Identifier(ForgeroStateRegistry.STATE_TO_CONTAINER.get(state.get().identifier()))))
-				.filter(state -> !Registry.ITEM.containsId(new Identifier(state.get().identifier())))
-				.sorted((element1, element2) -> compareStates(element1.get(), element2.get(), sortingMap))
-				.forEach(state -> {
-					try {
-						var converter = StateToItemConverter.of(state);
-						Identifier identifier = converter.id();
-						var item = converter.convert();
-						Registry.register(Registry.ITEM, identifier, item);
-					} catch (InvalidIdentifierException e) {
-						LOGGER.error("invalid identifier: {}", state.get().identifier());
-						LOGGER.error(e);
-					}
-				});
-	}
-
-	private int getOrderingFromState(Map<String, Integer> map, State state) {
-		var name = materialName(state);
-		int rarity = (int) state.stream().applyAttribute(AttributeType.RARITY);
-		return map.getOrDefault(name, rarity);
-	}
-
-	private String materialName(State state) {
-		var elements = state.name().split(ELEMENT_SEPARATOR);
-		if (elements.length > 1) {
-			return elements[0];
-		} else {
-			return state.name();
-		}
-	}
-
-	private int compareStates(State element1, State element2, Map<String, Integer> map) {
-		int elementOrdering = getOrderingFromState(map, element1) - getOrderingFromState(map, element2);
-		int nameOrdering = materialName(element1).compareTo(materialName(element2));
-
-		if (elementOrdering != 0) {
-			return elementOrdering;
-		} else if (nameOrdering != 0) {
-			return nameOrdering;
-		} else {
-			return rarity(element1) - rarity(element2);
-		}
-	}
-
-	private int rarity(State state) {
-		return (int) state.stream().applyAttribute(AttributeType.RARITY);
-	}
-
-	private void registerRecipes() {
-		RecipeRegistry.INSTANCE.registerRecipeSerializers();
+	private void handlePreInitializedEntryPoints() {
+		PRE_INITIALIZED_ENTRY_POINTS.forEach(ForgeroPreInitializationEntryPoint::onPreInitialization);
 	}
 }
